Argument count check done
"4. Actual and formal parameters number matching: when invoking a function, the number of
arguments passed via invocation must match the number of parameters that has been given
in the function definition. Otherwise, the error should be reported by the message: #lineno:
semantic error! Mismatch in numbers of arguments of 'ID', where 'ID' is the function that has
been invoked illegally."