Type checks
"7. Type mismatch: in a numerical and/or comparison operation, the types of operands on both
sides of the operation should match. Otherwise, the error should be reported by the message:
"#lineno: Semantic Error! Type mismatch in operands, Got 'Y' instead of 'X'", where 'Y' is the
mismatched type and 'X' is the expected type."